chore(blackjack): Move all source files to 'src' folder
